Удаляет компонент CartForm, а его данные выносит в основной компонент Cart.
